Improve repositores code
