+++ artifactory/commands/python/login.go
@@ -1,0 +1,64 @@
+package python
+
+import (
+	"github.com/jfrog/jfrog-cli-core/v2/artifactory/commands/repository"
+	cmdutils "github.com/jfrog/jfrog-cli-core/v2/artifactory/commands/utils"
+	"github.com/jfrog/jfrog-cli-core/v2/artifactory/utils"
+	"github.com/jfrog/jfrog-cli-core/v2/common/project"
+	"github.com/jfrog/jfrog-cli-core/v2/utils/config"
+	"github.com/jfrog/jfrog-cli-core/v2/utils/coreutils"
+	"github.com/jfrog/jfrog-client-go/artifactory/services"
+	"github.com/jfrog/jfrog-client-go/utils/log"
+)
+
+type PythonLoginCommand struct {
+	ר
+}
+
+func NewPythonLoginCommand() *PythonLoginCommand {
+	return &PythonLoginCommand{commandName: "rt_python_login"}
+}
+
+// Run configures python to use the specified or selected JFrog Artifactory repository
+// for package management, setting up registry and authentication.
+func (plc *PythonLoginCommand) Run() (err error) {
+	// If no repository is specified, prompt the user to select a pypi-compatible repository.
+	if plc.repo == "" {
+		// Define filter parameters to select virtual repositories of npm package type.
+		repoFilterParams := services.RepositoriesFilterParams{
+			RepoType:    utils.Virtual.String(),
+			PackageType: repository.Npm,
+		}
+
+		// Select repository interactively based on filter parameters and server details.
+		plc.repo, err = utils.SelectRepositoryInteractively(plc.serverDetails, repoFilterParams)
+		if err != nil {
+			return err
+		}
+	}
+
+	// Initialize NpmrcYarnrcManager for npm to manage registry and authentication configurations.
+	npmrcManager := cmdutils.NewNpmrcYarnrcManager(project.Npm, plc.repo, plc.serverDetails)
+
+	// Configure the registry URL for npm in the npm configuration.
+	if err = npmrcManager.ConfigureRegistry(); err != nil {
+		return err
+	}
+
+	// Configure authentication settings, handling token or basic auth as needed.
+	if err = npmrcManager.ConfigureAuth(); err != nil {
+		return err
+	}
+
+	// Output success message indicating successful npm configuration.
+	log.Output(coreutils.PrintTitle("Successfully configured npm client to work with your JFrog Artifactory repository: " + plc.repo))
+	return nil
+}
+
+func (plc *PythonLoginCommand) CommandName() string {
+	return plc.commandName
+}
+
+func (plc *PythonLoginCommand) ServerDetails() (*config.ServerDetails, error) {
+	return plc.serverDetails, nil
+}
